Added global high scores
The app now supports saving and loading high scores of all players. For this feature to function, internet connection is required. For now, the global high score record contains the player's amount of moves and the player's location (location services must be working and permitted for the latter to function). The records are stored on an external website.